keys update2

+++ src/utils.js
@@ -157,18 +157,54 @@ const bufferToStream = (buffer) => {
 	return Readable.from(buffer);
 };
 
-const reset = () => {
-	db.prepare('DELETE FROM users;').run();
-	db.prepare('DELETE FROM files;').run();
-	db.prepare('DELETE FROM parts;').run();
-	db.prepare('DELETE FROM nodes;').run();
+const reset = async () => {
+	const nodes = await getNodes(true, false, true);
+	if (nodes.length == 0) {
+		db.prepare('DELETE FROM users;').run();
+		db.prepare('DELETE FROM files;').run();
+		db.prepare('DELETE FROM parts;').run();
+		db.prepare('DELETE FROM nodes;').run();
 
-	db.prepare('DROP TABLE users;').run();
-	db.prepare('DROP TABLE files;').run();
-	db.prepare('DROP TABLE parts;').run();
-	db.prepare('DROP TABLE nodes;').run();
-	const { init } = require('./sql.js');
-	return init();
+		db.prepare('DROP TABLE users;').run();
+		db.prepare('DROP TABLE files;').run();
+		db.prepare('DROP TABLE parts;').run();
+		db.prepare('DROP TABLE nodes;').run();
+		const { init } = require('./sql.js');
+		return init();
+	} else {
+		for (let i = 0; i < nodes.length; i++) {
+			const node = nodes[i];
+
+			const body = {
+				key: node.ckey,
+			};
+
+			const agent = new Agent({
+				ca: node.ca,
+			});
+
+			await fetch(`https://${node.ip}:${node.port}/deinit`, {
+				method: 'POST',
+				body: JSON.stringify(body),
+				headers: { 'Content-Type': 'application/json' },
+				agent,
+			});
+
+			if (i == nodes.length - 1) {
+				db.prepare('DELETE FROM users;').run();
+				db.prepare('DELETE FROM files;').run();
+				db.prepare('DELETE FROM parts;').run();
+				db.prepare('DELETE FROM nodes;').run();
+
+				db.prepare('DROP TABLE users;').run();
+				db.prepare('DROP TABLE files;').run();
+				db.prepare('DROP TABLE parts;').run();
+				db.prepare('DROP TABLE nodes;').run();
+				const { init } = require('./sql.js');
+				return init();
+			}
+		}
+	}
 };
 
 module.exports = {
